Reorganize, add theming for cover cards
