+++ client/src/index.css
@@ -188,7 +188,7 @@ code {
 }
 
 .wiki-menu-colors {
-    --background-color: #2a2518;
+    --background-color: #282317;
     --primary-color: #1c160d;
     --primary-border-color: #52402e;
     --secondary-color: #362c24;
