+++ client/src/menu/Settings.tsx
@@ -1,4 +1,4 @@
-import React, { JSXElementConstructor, ReactElement, useEffect, useState } from 'react';
+import React, { JSXElementConstructor } from 'react';
 import "./settings.css";
 import translate, { LANGUAGES, Language, languageName, switchLanguage } from '../game/lang';
 import GAME_MANAGER from '..';
@@ -6,12 +6,10 @@ import Anchor from './Anchor';
 import StartMenu from './main/StartMenu';
 import LoadingScreen from './LoadingScreen';
 import { loadSettings, saveSettings } from '../game/localStorage';
-import GameModesEditor from './GameModesEditor';
+import GameModesEditor from '../components/GameModesEditor';
 import { CopyButton } from '../components/ClipboardButtons';
-import Wiki from '../components/Wiki';
-import { Role } from '../game/roleState.d';
-import { StateListener } from '../game/gameManager.d';
 import ReactDOM from 'react-dom';
+import WikiCoverCard from '../components/WikiCoverCard';
 
 export type Settings = {
     volume: number,
@@ -122,42 +120,12 @@ export default class SettingsMenu extends React.Component<SettingsProps, Setting
                     localStorage.clear();
                 }}>{translate('menu.settings.eraseSaveData')}</button>
                 <button onClick={() => {
-                    Anchor.setCoverCard(<WikiCoverCard />);
+                    Anchor.setCoverCard(<WikiCoverCard />, "wiki-menu-colors");
                     Anchor.closeSettings();
                 }}>{translate("menu.wiki.title")}</button>
             </div>
         );
     }
-}
-
-function WikiCoverCard(): ReactElement {
-    let defaultDisabledRoles: Role[];
-    switch (GAME_MANAGER.state.stateType) {
-        case "disconnected":
-        case "outsideLobby":
-            defaultDisabledRoles = [];
-        break;
-        case "game":
-        case "lobby":
-            defaultDisabledRoles = GAME_MANAGER.state.excludedRoles;
-        break;
-    }
-    const [disabledRoles, setDisabledRoles] = useState(defaultDisabledRoles);
-
-    useEffect(() => {
-        const listener: StateListener = (type) => {
-            if (type === "excludedRoles" && (GAME_MANAGER.state.stateType === "game" || GAME_MANAGER.state.stateType === "lobby")) {
-                setDisabledRoles(GAME_MANAGER.state.excludedRoles)
-            }
-        }
-        GAME_MANAGER.addStateListener(listener);
-        return () => GAME_MANAGER.removeStateListener(listener);
-    }, []);
-
-    return <div className='wiki-cover-card'>
-        <h1>{translate("menu.wiki.title")}</h1>
-        <Wiki disabledRoles={disabledRoles}/>
-    </div>
 }
 
 export function RoomCodeButton(): JSX.Element {
