+++ client/src/menu/anchor.css
@@ -123,5 +123,5 @@
     width: 100%;
     height: 100%;
     background-color: var(--primary-color);
-    opacity: 80%;
+    opacity: 60%;
 }
